+++ NEWS.md
@@ -87,7 +87,7 @@ This is a small release focusing on fixing regressions from 3.5.0 and
   @yutannihilation's prior work, #3120)
 * When discrete breaks have names, they'll be used as labels by default 
   (@teunbrand, #6147).
-* The helper function `is.waiver()` is now exported to help extensions to work
+* The helper function `is_waiver()` is now exported to help extensions to work
   with `waiver()` objects (@arcresu, #6173).
 * Date(time) scales now throw appropriate errors when `date_breaks`, 
   `date_minor_breaks` or `date_labels` are not strings (@RodDalBen, #5880)